requirements.txt
python-telegram-bot==20.7
web3==6.15.1
python-dotenv==1.0.0
requests==2.31.0
asyncio
